Get finer precision text extent
On the Mac port, the cursor tends to "drift" in longer lines, so that it
does not correclty highlight the corresponding character towards the end
of the line.

This is because the text extent is returned as a rounded integer. The
error in this rounding multiples over longer lines, causing this drift.

This change finds the extent for 10 characters, rather than 1, and then
divides by 10.0, to get a float extent. This eliminates the drift.

When drawing the cursor, we then round this extent back up to avoid the
cursor appearing too narrow.